requirements.txt
flask==3.0.3
requests==2.32.3
gspread==6.1.2
oauth2client==4.1.3
psycopg2-binary==2.9.9
redis==5.0.8
gunicorn==22.0.0